+++ appcenter-analytics/Analytics.js
@@ -12,6 +12,11 @@ const Analytics = {
     }),
 
     // async - returns a Promise
+    startSession() {
+        return AppCenterReactNativeAnalytics.startSession();
+    },
+
+    // async - returns a Promise
     trackEvent(eventName, properties) {
         return AppCenterReactNativeAnalytics.trackEvent(eventName, sanitizeProperties(properties));
     },